Create Tic Tac Toe
Tic-tac-toe is a paper-and-pencil game for two players, X and O, who take turns marking the spaces in a 3×3 grid.The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal row wins the game. Here, we have developed a mini project in C Tic Tac Toe game – a simple console application without graphics. It is the same noughts and crosses or the Xs and Os, the other names for Tic-Tac-Toe, we have played with paper and pencil during our school days! This Tic Tac Toe game in C is compiled in Code::Blocks with gcc compiler.

Developed by Sanika Saini and Abhimanyu Dudeja (RA2111003010971 AND RA2111003010989)